+++ Templates/index.html
@@ -18,9 +18,11 @@
     <link href="https://fonts.googleapis.com/css?family=Lato" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css?family=Catamaran:100,200,300,400,500,600,700,800,900" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css?family=Muli" rel="stylesheet">
+  <!-- d3 JavaScript -->
+  <script src="https://d3js.org/d3.v5.min.js"></script>
 
     <!-- Custom styles for this template -->
-    <link href="css/style.min.css" rel="stylesheet">
+  <link href="../static/css/style.min.css" rel="stylesheet" type="text/css">
 
   </head>
 
@@ -55,8 +57,8 @@
         <div class="row h-100">
           <div class="col-lg-7 my-auto">
             <div class="header-content mx-auto">
-              <h1 class="mb-5">RNN Training and Sentiment Analysis</h1>
-              <p>What is the project about?  You need a summary here but nothing too long.</p>
+              <h1 class="mb-5">Can a Machine Learn Sentiment?</h1>
+              <p></p>
             </div>
           </div>
           <div class="col-lg-5 my-auto">
@@ -126,8 +128,8 @@
     <!-- Plugin JavaScript -->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-easing/1.4.1/jquery.easing.min.js"></script>
 
-    <!-- Custom scripts for this template -->
-    <script src="js/main.min.js"></script>
+    <script src= "../static/js/main.min.js"></script>
+
 
   </body>
