+++ app.py
@@ -1,93 +1,40 @@
+import sys
 import os
 import io
 import numpy as np
+from flask import Flask, request, redirect, url_for, jsonify, render_template
 
-import keras
-from keras.preprocessing import image
-from keras.preprocessing.image import img_to_array
-from keras.applications.xception import (
-    Xception, preprocess_input, decode_predictions)
 from keras import backend as K
-
-from flask import Flask, request, redirect, url_for, jsonify
+from keras.models import Sequential
+from keras.layers import Dense
+from keras.layers import Dropout
+from keras.layers import LSTM
+from keras.callbacks import ModelCheckpoint
+from keras.utils import np_utils
 
 app = Flask(__name__)
 app.config['UPLOAD_FOLDER'] = 'Uploads'
 
-model = None
-graph = None
+@app.route("/")
+def index():
+    """Return the homepage."""
+    return render_template("index.html")
 
+# @app.route('/', methods=['GET', 'POST'])
+# def upload_file():
+#     data = {"success": False}
+#     if request.method == 'POST':
+#         if request.files.get('file'):
+#             # read the file
+#             file = request.files['file']
 
-def load_model():
-    global model
-    global graph
-    model = Xception(weights="imagenet")
-    graph = K.get_session().graph
+#             # read the filename
+#             filename = file.filename
 
+#             # create a path to the uploads folder
+#             filepath = os.path.join(app.config['UPLOAD_FOLDER'], filename)
 
-load_model()
-
-
-def prepare_image(img):
-    img = img_to_array(img)
-    img = np.expand_dims(img, axis=0)
-    img = preprocess_input(img)
-    # return the processed image
-    return img
-
-
-@app.route('/', methods=['GET', 'POST'])
-def upload_file():
-    data = {"success": False}
-    if request.method == 'POST':
-        if request.files.get('file'):
-            # read the file
-            file = request.files['file']
-
-            # read the filename
-            filename = file.filename
-
-            # create a path to the uploads folder
-            filepath = os.path.join(app.config['UPLOAD_FOLDER'], filename)
-
-            file.save(filepath)
-
-            # Load the saved image using Keras and resize it to the Xception
-            # format of 299x299 pixels
-            image_size = (299, 299)
-            im = keras.preprocessing.image.load_img(filepath,
-                                                    target_size=image_size,
-                                                    grayscale=False)
-
-            # preprocess the image and prepare it for classification
-            image = prepare_image(im)
-
-            global graph
-            with graph.as_default():
-                preds = model.predict(image)
-                results = decode_predictions(preds)
-                data["predictions"] = []
-
-                # loop over the results and add them to the list of
-                # returned predictions
-                for (imagenetID, label, prob) in results[0]:
-                    r = {"label": label, "probability": float(prob)}
-                    data["predictions"].append(r)
-
-                # indicate that the request was a success
-                data["success"] = True
-
-        return jsonify(data)
-
-    return '''
-    <!doctype html>
-    <title>Upload new File</title>
-    <h1>Upload new File</h1>
-    <form method=post enctype=multipart/form-data>
-      <p><input type=file name=file>
-         <input type=submit value=Upload>
-    </form>
-    '''
+#             file.save(filepath)
 
 
 if __name__ == "__main__":
