+++ static/scss/new-age.scss
@@ -1,0 +1,11 @@
+@import "variables.scss";
+@import "mixins.scss";
+@import "global.scss";
+@import "navbar.scss";
+@import "masthead.scss";
+@import "download.scss";
+@import "features.scss";
+@import "cta.scss";
+@import "contact.scss";
+@import "footer.scss";
+@import "bootstrap-overrides.scss";
